Move test asserting firing container lifecycle payloads into Full group
